AI_Agents/Pizza_AI_Agent/Pizza_AI_Agent/requirements.txt
flask
requests
markdown
python-dotenv
pandas
openai
langchain-core
langchain
langchain-community
langchainhub
sentence-transformers
chromadb
gunicorn
langchain_ollama
langchain_chroma
langchain-huggingface
